# Remove unncessecary allocation in WAL.Write(…)